python: move modifiers into own compilation units
